requirements.txt
scikit-image==0.26.0
matplotlib==3.10.8
python-dotenv==1.2.2
scikit-learn==1.8.0
torch==2.11.0
pandas==3.0.1
imageio==2.37.2
ipywidgets==8.1.8
pillow==12.1.1
opencv-python==4.13.0.92
huggingface_hub==1.5.0
numpy==2.4.2
kagglehub==1.0.0
hydra-core==1.3.2
torchvision==0.26.0
iopath==0.1.10
pycocotools==2.0.11
antspyx==0.6.3
scipy==1.17.1
attrs==26.1.0